API/WBT/court-auction-crawler-0.1.3/crawler/pdftool/utils/uitls.py
import re
from dateutil import parser


def remove_blank(text):
    return str(text).strip().replace('\r', '').replace('\t', '').replace('\n', '').replace(u'\u3000', u'').replace(u'\xa0', u'').replace(' ', '')
